+++ internal/distro/fedora30/ext4.go
@@ -1,10 +1,13 @@
-package blueprint
+package fedora30
 
-import "github.com/osbuild/osbuild-composer/internal/pipeline"
+import (
+	"github.com/osbuild/osbuild-composer/internal/blueprint"
+	"github.com/osbuild/osbuild-composer/internal/pipeline"
+)
 
 type ext4Output struct{}
 
-func (t *ext4Output) translate(b *Blueprint) (*pipeline.Pipeline, error) {
+func (t *ext4Output) translate(b *blueprint.Blueprint) (*pipeline.Pipeline, error) {
 	packages := [...]string{
 		"policycoreutils",
 		"selinux-policy-targeted",
@@ -18,13 +21,13 @@ func (t *ext4Output) translate(b *Blueprint) (*pipeline.Pipeline, error) {
 	}
 	p := getCustomF30PackageSet(packages[:], excludedPackages[:], b)
 	addF30LocaleStage(p)
-	addF30GRUB2Stage(p, b.getKernelCustomization())
+	addF30GRUB2Stage(p, b.GetKernelCustomization())
 	addF30FixBlsStage(p)
 	addF30SELinuxStage(p)
 	addF30RawFSAssembler(p, t.getName())
 
 	if b.Customizations != nil {
-		err := b.Customizations.customizeAll(p)
+		err := customizeAll(p, b.Customizations)
 		if err != nil {
 			return nil, err
 		}
