+++ internal/distro/fedora30/openstack.go
@@ -1,10 +1,13 @@
-package blueprint
+package fedora30
 
-import "github.com/osbuild/osbuild-composer/internal/pipeline"
+import (
+	"github.com/osbuild/osbuild-composer/internal/blueprint"
+	"github.com/osbuild/osbuild-composer/internal/pipeline"
+)
 
 type openstackOutput struct{}
 
-func (t *openstackOutput) translate(b *Blueprint) (*pipeline.Pipeline, error) {
+func (t *openstackOutput) translate(b *blueprint.Blueprint) (*pipeline.Pipeline, error) {
 	packages := [...]string{
 		"@Core",
 		"chrony",
@@ -24,13 +27,13 @@ func (t *openstackOutput) translate(b *Blueprint) (*pipeline.Pipeline, error) {
 	p := getCustomF30PackageSet(packages[:], excludedPackages[:], b)
 	addF30LocaleStage(p)
 	addF30FSTabStage(p)
-	addF30GRUB2Stage(p, b.getKernelCustomization())
+	addF30GRUB2Stage(p, b.GetKernelCustomization())
 	addF30FixBlsStage(p)
 	addF30SELinuxStage(p)
 	addF30QemuAssembler(p, "qcow2", t.getName())
 
 	if b.Customizations != nil {
-		err := b.Customizations.customizeAll(p)
+		err := customizeAll(p, b.Customizations)
 		if err != nil {
 			return nil, err
 		}
