+++ test/cases/disk_empty_blueprint.json
@@ -1,5 +1,6 @@
 {
   "compose": {
+    "distro": "fedora-30",
     "output-format": "partitioned-disk",
     "filename": "disk.img",
     "blueprint": {}
