+++ test/cases/openstack_empty_blueprint.json
@@ -1,5 +1,6 @@
 {
   "compose": {
+    "distro": "fedora-30",
     "output-format": "openstack",
     "filename": "image.qcow2",
     "blueprint": {}
