+++ test/cases/vhd_empty_blueprint.json
@@ -1,5 +1,6 @@
 {
   "compose": {
+    "distro": "fedora-30",
     "output-format": "vhd",
     "filename": "image.vhd",
     "blueprint": {}
